Split protobufs into logical namespaces
This is not only cleaner, it allows us to do smart things around exporting certain rpcs to different locations, should we need to.
