+++ messaging.go
@@ -7,7 +7,7 @@ import (
 	"google.golang.org/protobuf/types/known/emptypb"
 )
 
-func (s Server) Subscribe(_ *emptypb.Empty, ss server.Proxy_SubscribeServer) (err error) {
+func (s Server) Subscribe(_ *emptypb.Empty, ss server.Messaging_SubscribeServer) (err error) {
 	u, err := s.userFromContext(ss.Context())
 	if err != nil {
 		return
@@ -55,12 +55,44 @@ func (s Server) Send(ctx context.Context, in *server.MessageWrapper) (out *empty
 	return
 }
 
-func contains(sl []string, s string) bool {
-	for _, elem := range sl {
-		if elem == s {
-			return true
+func (s Server) PublicKey(in *server.Auth, pks server.Messaging_PublicKeyServer) (err error) {
+	u, err := s.userFromContext(pks.Context())
+	if err != nil {
+		return
+	}
+
+	// todo: this is the point at which we need to test the hostname of
+	// in.UserId - if the hostname is our host then we can lookup in redis.
+	//
+	// else: we need to query the remote host associated with this hostname
+	// and get public keys from there
+
+	lookupIds := make([]string, 0)
+	if groupRegexp.Match([]byte(in.UserId)) {
+		g, err := s.redis.Group(in.UserId)
+		if err != nil || !HasPermission(u, g, groupRead) {
+			return badGroupError
+		}
+
+		for _, u := range g.Members {
+			lookupIds = append(lookupIds, u)
+		}
+	} else {
+		lookupIds = append(lookupIds, in.UserId)
+	}
+
+	for _, id := range lookupIds {
+		keys, err := s.redis.GetPublicKeys(id)
+		if err != nil {
+			return generalError
+		}
+
+		for _, key := range keys {
+			pks.Send(&server.PublicKeyValue{
+				Value: key,
+			})
 		}
 	}
 
-	return false
+	return
 }
